Towny Problem
As you are reading this, PLEASE keep in mind that I changed my name BEFORE you posted that warning on the front page!

So, as I said, I changed my name, now my old IGN is the mayor of Jerusalem, and I'm not sure what to do. But what I can assume is that Towny still runs off the old Usernames system, so my solution would be to add me to the town, and make me mayor. Not sure what you guys would do about it though.